+++ package.json
@@ -34,7 +34,8 @@
     "react-native-fs": "^2.18.0",
     "reflect-metadata": "^0.1.13",
     "tsyringe": "^4.5.0",
-    "uuid": "^8.3.0"
+    "uuid": "^8.3.0",
+    "ws": "^7.4.5"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
@@ -63,8 +64,7 @@
     "ts-jest": "^26.5.3",
     "ts-node-dev": "^1.1.6",
     "tslog": "^3.1.2",
-    "typescript": "^4.2.3",
-    "ws": "^7.4.5"
+    "typescript": "^4.2.3"
   },
   "publishConfig": {
     "access": "public",
