+++ src/transport/OutboundTransporter.ts
@@ -4,4 +4,7 @@ export interface OutboundTransporter {
   supportedSchemes: string[]
 
   sendMessage(outboundPackage: OutboundPackage): Promise<any>
+
+  start(): Promise<void>
+  stop(): Promise<void>
 }
