+++ src/transport/index.ts
@@ -1,3 +1,4 @@
 export * from './InboundTransporter'
 export * from './OutboundTransporter'
 export * from './HttpOutboundTransporter'
+export * from './WsOutboundTransporter'
